Añadir un mensaje para que confirme que quiere eliminar #8

+++ src/main/java/controller/ControllerImplementation.java
@@ -129,7 +129,7 @@ public class ControllerImplementation implements IController, ActionListener {
             case Constants.STORAGE_FILE:
                 setupFileStorage();
                 break;
-            case Constants.STORAGE_FILE_SERIALIZATION :
+            case Constants.STORAGE_FILE_SERIALIZATION:
                 setupFileSerialization();
                 break;
             case Constants.STORAGE_SQL:
@@ -439,7 +439,12 @@ public class ControllerImplementation implements IController, ActionListener {
     public void delete(Person p) {
         try {
             if (dao.read(p) != null) {
-                dao.delete(p);
+                int opc = JOptionPane.showConfirmDialog(delete, "Are you sure you want to delete this person?", delete.getTitle(), JOptionPane.YES_NO_OPTION);
+                if (opc == javax.swing.JOptionPane.YES_NO_OPTION) {
+                    dao.delete(p);
+                    JOptionPane.showMessageDialog(delete, "Person deleted successfully!", delete.getTitle(), JOptionPane.INFORMATION_MESSAGE);
+                }
+
             } else {
                 throw new PersonException(p.getNif() + " is not registered and can not "
                         + "be DELETED");
